harness-doctor: bring audit.py to parity with audit.ts
The Python side computed a different summary from the same tables: its own
markdown layout, its own ranking shapes. Splits the aggregates into
aggregate_commands, aggregate_human, and aggregate_projects to mirror the
TypeScript modules, so both runtimes now emit byte-identical fact tables,
JSON, and markdown over one frozen corpus.

Documents the actual `--format json` shape in stats.md: every ranking is
{rows, shown, total, dropped}, byModel is a proportional attribution rather
than a measurement, and the flags get a table of their own.

+++ skills/inspect-and-debug/harness-doctor/scripts/lib/types.py
@@ -22,8 +22,8 @@ class SessionRow(TypedDict, total=False):
     cwd: str
     branch: str
     ver: str
-    start: float
-    end: float
+    start: int
+    end: int
     kind: SessionKind
     parent: str
     run: str
@@ -47,7 +47,7 @@ class ToolRow(TypedDict, total=False):
     sid: str
     proj: str
     kind: SessionKind
-    t: float
+    t: int
     tool: str
     durSec: float
     err: bool
@@ -62,7 +62,7 @@ class BashRow(TypedDict, total=False):
     sid: str
     proj: str
     kind: SessionKind
-    t: float
+    t: int
     durSec: float
     err: bool
     outChars: int
@@ -81,7 +81,7 @@ class UserMsgRow(TypedDict, total=False):
     sid: str
     proj: str
     kind: SessionKind
-    t: float
+    t: int
     text: str
     side: bool
     human: bool
@@ -93,7 +93,7 @@ class ErrorRow(TypedDict, total=False):
     sid: str
     proj: str
     kind: SessionKind
-    t: float
+    t: int
     tool: str
     input: str
     msg: str
